% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcglm.R
\name{bcglm}
\alias{bcglm}
\title{Bayesian Compositional GLMs}
\usage{
bcglm(x, y, family = gaussian, plot = TRUE)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (binary or continuous)}

\item{family}{gaussian or bernoulli}

\item{plot}{if TRUE, shows the plots (default = TRUE)}
}
\value{

}
\description{
Bayesian Compositional GLMs
}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
\author{
Li Zhang
}
