% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcglmm.R
\name{bcglmm}
\alias{bcglmm}
\title{Bayesian Compositional GLMMs}
\usage{
bcglmm(x, y, family = gaussian, dist = "bray")
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (binary or continuous)}

\item{family}{gaussian or bernoulli}

\item{dist}{method to calculate dissimilarity matrix, default is Bray-curtis}
}
\value{

}
\description{
Bayesian compositional GLMMs methods for microbiome data with continuous or binary outcome.
}
\examples{
dat=sim_cmm(200,100,0.2)
          fit=bcglmm(dat$x,dat$y)
          summary(fit)
          fixef(fit)
          mcmc_plot(fit,variable = "^b_X", regex = TRUE)
          plot(fit,variable=c("b_X18","b_X20","b_X22","b_X24","b_X26","b_X28"),N = 6)

}
\references{
\url{https://link.springer.com/article/10.1186/s12859-025-06114-3}
}
\author{
Li Zhang
}
