% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bco.R
\name{bco}
\alias{bco}
\title{Bayesian Compositional Ordinal Regression}
\usage{
bco(x, y, plot = TRUE)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (ordinal)}

\item{plot}{if TRUE, shows the plots (default = TRUE)}
}
\value{

}
\description{
Bayesian Compositional Ordinal Regression
}
\references{
\url{https://journals.sagepub.com/doi/abs/10.1177/09622802241247730}
}
\author{
Li Zhang
}
