% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_b.R
\name{sim_b}
\alias{sim_b}
\title{Simulation for binary response and compositional predictors}
\usage{
sim_b(n, p, q = 0.5)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension of compositional predictors}

\item{q}{quantile for generating binary response}
}
\value{

}
\description{
Simulation of data with dependent compositional predictors and binary response.
The simulation setting and dependence structure can be inferred from paper.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
