% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cmm.R
\name{sim_cmm}
\alias{sim_cmm}
\title{Simulation for continuous response and mixed effects compositional predictors}
\usage{
sim_cmm(n, p, snr)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension of compositional predictors}

\item{snr}{proportion of small effects to the total number of predictors p}
}
\value{

}
\description{
Simulation of data with dependent compositional predictors and continuous response.
The simulation setting and dependence structure can be inferred from paper.
}
\references{
\url{https://link.springer.com/article/10.1186/s12859-025-06114-3}
}
