% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_o.R
\name{sim_o}
\alias{sim_o}
\title{Simulation for ordinal response and compositional predictors}
\usage{
sim_o(n, p, q = c(0.2, 0.5))
}
\arguments{
\item{n}{sample size}

\item{p}{dimension of compositional predictors}

\item{q}{quantiles for generating ordinal response}
}
\value{

}
\description{
Simulation for ordinal response and compositional predictors
}
