% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_s.R
\name{sim_s}
\alias{sim_s}
\title{Simulation for survival response and compositional predictors}
\usage{
sim_s(n, p)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension of compositional predictors}
}
\value{

}
\description{
Simulation of data with dependent compositional predictors and survival response.
The simulation setting and dependence structure can be inferred from paper.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
