% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Similarity matrix calculation}
\usage{
similarity(dis.taxa)
}
\arguments{
\item{dis.taxa}{distance matrix calculated based on either phylogenetic or non-phylogenetic methods}
}
\value{

}
\description{
We provide phylogenetic similarity choice to capture the phylogenetic relatedness among taxa.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
\author{
Li Zhang
}
