% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccoxph.R
\name{bccoxph}
\alias{bccoxph}
\title{Bayesian Compositional survival analysis}
\usage{
bccoxph(x, y, df_local = 1, df_global = 1, similarity = NULL)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{survival outcome}

\item{df_local}{the degree of freedom in half-t prior for local scale in horseshoe prior (default=1)}

\item{df_global}{the degree of freedom in half-t prior for global scale in horseshoe prior (default=1)}

\item{similarity}{measures the relatedness among taxa (default=NULL)}
}
\value{

}
\description{
Bayesian compositional survival analysis for microbiome data.
}
\examples{
dat=sim_s(n=400,p=100)
          x=otu_table(dat$x,taxa_are_rows = F)
          dis.taxa = phyloseq::distance(x, method = "bray", type = "taxa")
          sim=similarity(dis.taxa)
          fit=bccoxph(x=dat$x,y=dat$y,df_local=1,df_global=1,similarity=sim)
          summary(fit)
          fixef(fit)
          mcmc_plot(fit,variable = "^b_x", regex = TRUE)
          plot(fit,variable=c("b_x18","b_x20","b_x22","b_x24","b_x26","b_x28"),nvariables = 6)

}
\author{
Zhenying Ding
}
