% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcglm.R
\name{bcglm}
\alias{bcglm}
\title{Bayesian Compositional GLMs}
\usage{
bcglm(x, y, family = gaussian, df_local = 1, df_global = 1, similarity = NULL)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (binary or continuous)}

\item{family}{gaussian or bernoulli}

\item{df_local}{the degree of freedom in half-t prior for local scale in horseshoe prior (default=1)}

\item{df_global}{the degree of freedom in half-t prior for global scale in horseshoe prior (default=1)}

\item{similarity}{measures the relatedness among taxa (default=NULL)}
}
\value{

}
\description{
Bayesian compositional GLMs methods for microbiome data with continuous or binary outcome.
}
\examples{
dat=sim_c(200,100)
          sim=similarity(dat$x)
          fit=bcglm(x=dat$x,y=dat$y,family=gaussian,df_local=1,df_global=1,similarity=sim)
          summary(fit)
          fixef(fit)
          mcmc_plot(fit,variable = "^b_X", regex = TRUE)
          plot(fit,variable=c("b_X18","b_X20","b_X22","b_X24","b_X26","b_X28"),nvariables = 6)

}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
\author{
Li Zhang
}
