% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcglm.R
\name{bcglm}
\alias{bcglm}
\title{Bayesian Compositional GLMs}
\usage{
bcglm(x, y, family = gaussian, similarity = NULL)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (binary or continuous)}

\item{family}{gaussian or bernoulli}
}
\value{

}
\description{
Bayesian compositional GLMs methods for microbiome data with continuous or binary outcome.
}
\examples{
dat=sim_c(200,100)
          sim=similarity(dat$x)
          fit=bcglm(dat$x,dat$y,similarity=sim)
          summary(fit)
          fixef(fit)
          mcmc_plot(fit)

}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
\author{
Li Zhang
}
