% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcglmm.R
\name{bcglmm}
\alias{bcglmm}
\title{Bayesian Compositional GLMMs}
\usage{
bcglmm(
  x,
  y,
  otu,
  family = gaussian,
  df_local = 1,
  df_global = 1,
  dist = "bray"
)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (binary or continuous)}

\item{otu}{a phyloseq object or otu_table object that used to calculate distance matrix}

\item{family}{gaussian or bernoulli}

\item{df_local}{the degree of freedom in half-t prior for local scale in horseshoe prior (default=1)}

\item{df_global}{the degree of freedom in half-t prior for global scale in horseshoe prior (default=1)}

\item{dist}{method to calculate dissimilarity matrix, default is Bray-curtis}
}
\value{

}
\description{
Bayesian compositional GLMMs methods for microbiome data with continuous or binary outcome.
}
\examples{
dat=sim_cmm(n=400,p=100,snr=0.2)
          otu=otu_table(dat$x,taxa_are_rows = F)
          fit=bcglmm(x=dat$x,y=dat$y,otu=otu,family=gaussian,df_local=3,df_global=3,dist="bray")
          summary(fit)
          fixef(fit)
          mcmc_plot(fit,variable = "^b_X", regex = TRUE)
          plot(fit,variable=c("b_X18","b_X20","b_X22","b_X24","b_X26","b_X28"),nvariables = 6)

}
\references{
\url{https://link.springer.com/article/10.1186/s12859-025-06114-3}
}
\author{
Li Zhang
}
