% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bco.R
\name{bco}
\alias{bco}
\title{Bayesian Compositional Ordinal Regression}
\usage{
bco(x, y, similarity = NULL)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (ordinal)}

\item{similarity}{measures the relatedness among taxa (deafult=NULL)}
}
\value{

}
\description{
Bayesian compositional method for microbiome data with ordinal outcome.
}
\examples{
dat=sim_o(200,100)
          fit=bco(dat$x,dat$y)
          summary(fit)
          fixef(fit)
          mcmc_plot(fit,variable = "^b_X", regex = TRUE)
          plot(fit,variable=c("b_X18","b_X20","b_X22","b_X24","b_X26","b_X28"),N = 6)

}
\references{
\url{https://journals.sagepub.com/doi/abs/10.1177/09622802241247730}
}
\author{
Li Zhang
}
