% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bco.R
\name{bco}
\alias{bco}
\title{Bayesian Compositional Ordinal Regression}
\usage{
bco(x, y, df_local = 1, df_global = 1, similarity = NULL)
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{y}{outcome (ordinal)}

\item{df_local}{the degree of freedom in half-t prior for local scale in horseshoe prior (default=1)}

\item{df_global}{the degree of freedom in half-t prior for global scale in horseshoe prior (default=1)}

\item{similarity}{measures the relatedness among taxa (default=NULL)}
}
\value{

}
\description{
Bayesian compositional method for microbiome data with ordinal outcome.
}
\examples{
dat=sim_o(n=400,p=100,q=c(0.2,0.5))
          fit=bco(x=dat$x,y=dat$y,df_local=1,df_global=1)
          summary(fit)
          fixef(fit)
          mcmc_plot(fit,variable = "^b_X", regex = TRUE)
          plot(fit,variable=c("b_X18","b_X20","b_X22","b_X24","b_X26","b_X28"),nvariables = 6)

}
\references{
\url{https://journals.sagepub.com/doi/abs/10.1177/09622802241247730}
}
\author{
Li Zhang
}
