% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_c.R
\name{sim_c}
\alias{sim_c}
\title{Simulation for continuous response and compositional predictors}
\usage{
sim_c(n, p)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension of compositional predictors}
}
\value{

}
\description{
Simulation of data with dependent compositional predictors and continuous response.
The simulation setting and dependence structure can be inferred from paper.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
