% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Similarity matrix calculation}
\usage{
similarity(x, dist = "bray")
}
\arguments{
\item{x}{abundance matrix or data frame (rows are samples, columns are variables (taxa))}

\item{dist}{method to calculate dissimilarity matrix, default is Bray-curtis}
}
\value{

}
\description{
We provide phylogenetic similarity choice to capture the phylogenetic relatedness among taxa.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.9946}
}
\author{
Li Zhang
}
